I run some benchmarks for comparison, basically I cloned `TestGenMap` test and changed it for benchmarking. Here are the results
### Before 
go test -benchmem -run=github.com/leanovate/gopter -bench BenchmarkMap
goos: darwin
goarch: amd64
pkg: github.com/leanovate/gopter
BenchmarkMap-8             **30000             54663 ns/op           28240 B/op         34 allocs/op**
PASS
ok      github.com/leanovate/gopter     2.183s
### After
go test -benchmem -run=github.com/leanovate/gopter -bench BenchmarkMap
goos: darwin
goarch: amd64
pkg: github.com/leanovate/gopter
BenchmarkMap-8            **100000             22058 ns/op           11872 B/op         25 allocs/op**
PASS
ok      github.com/leanovate/gopter     2.451s

This is a PR for #29